Added plev_flag
Kept getting "ValueError: The truth value of an array with more than one element is ambiguous” error when using the
      if plev:
construct with a vector. Hard to think of an easy way to distinguish between None and [x1, x2, ..., xn] so I just added another variable plev_flag = 1 for fields with pressure coords and =0 for surface variables.