Refactor: empty array generator
Added "modulo" option.

For the fetch_nc_data function it makes sense to preallocate an empty array that is of a size that can be reshaped into [months years ...] shape (faster to preallocate the padding after the fact).

Currently the empty array generator has no way of doing this (always generates an array of size time_length).

Modulo arguement is optional (default None) takes either None or integer.  If an integer it will round the the size of the time_dimension argument such that time_argument/modulo has 0 remainder (ie can be correctly reshaped into size [modulo, years, ...]. If modulo = None it is set to = time_length and uses the fact that N%N = 0 for any N.

Should,'t have any side effects as if Modulo=None the array generated will have size time_length as always