Bugfix: empty_array_generator
empty_array_generator has an optional arg model_dim. This arg creates a 4th dimension (notionally to store different model  runs in a big numpy array). Previously model_dim=1 was the default and np.squeeze was used to remove the dimension, giving a 3d array
           [time,plev,lat,1] ==> [time_plev,lat]
However this technique results in an error if we actually fdo want model_dim = 1. Replaced with model_dim=None and the use of an if statement.