empty_array_generator refactored
Instead of passing files and the function getting plev and lat and then having various options for additional variables (model, experiment, variables etc.) now just pass a list of numbers which become the array dimenison lengths.

Had to change the absense of pressure from being plev= None to plev = [] because len(None) is not defined but len([])=0

Moved the modulo padding that appeared in the orignal empty_generator_function to its own function.

This is a far more flexible and  useful construction