New idealised model data handling scripts
This script is somewhat similar to nc_data.py and output_nc_data.py but written for the idealised model output. There is a bit of code duplication here but the model output is so much simpler (and directory structure quite different) I think it is worth just having a seperate idealised model py script.

I'm not sure whether I should put it in a seperate directoyr or repo but will put it here for now