Refactored code: defined get_time_dim
I have decided to package a code chunk that appears in output_nc_data.py as a seperate function.

This function sweeps through a list of /experi/freq/realm/vari/models and finds the max/min time length to be used when preallocating an array.

I have found a need for this function when calculating the area-weighted surface temperature so it makes since to atomise it.