No longer squeeze output array from fetch_nc_data
I have made the desicion to longer squeeze the array (and collapse any length = 1 dimensions.

This is because in some of the climate variable calculation I want to be able to do loop over model and experiment. Sometimes these will only have length 1 and if we squeeze the resulting output array, a varying number of array dimensions will result.