Fixed a bug in fetch nc_data
The squeeze() shouln't be there. The HadGEM model has a nc file with only 1 time value. When this is squeezed the 0th dim becomes lat dimension and it creates problems (but not the prgram to fail!). Basically the arrays still broadcast to the right shape and you get ~30 months of that one time value created.